pyproject.toml
[project]
name = "testjwt"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "pyjwt>=2.10.1",
]
